requirements.txt
pandas
requests
certifi
openpyxl
mysql-connector-python
python-dotenv
googlemaps